--- packages/config/scripts/omnichain/types.ts
@@ -1,9 +1,0 @@
-export interface LZData {
-  address: string
-  deployer: string
-  name: string
-  count: number
-  tvl: string
-  ethValue: string
-  isERC20: string
-}
